Decompose integrations/registration.go God Object (824→49 LOC)
Extract 6 self-contained registerX functions out of the 824-line
registration.go God Object into their domain homes:

- ado.Register()          → ado/registration.go      (528 LOC)
- atlassian.RegisterConfluence() → atlassian/register.go
- atlassian.RegisterJira()      → atlassian/register.go
- registerMedia()         → media.go                 (same package)
- registerNetwork()       → network.go               (same package)
- registerTeams()         → teams.go                 (same package)

registration.go retains only RegisterAll orchestrator (49 LOC,
complexity 1) that calls each domain's Register in fail-fast
sequence with preserved call order.

Architecture:
- ADR-028 documents the sub-package registration pattern
- Zero behavioral changes — pure mechanical extraction
- Import graph: acyclic parent→child
- verify_architecture: clean
- go test -race ./internal/tools/integrations/...: pass

Closes #116
