+++ internal/tools/integrations/network.go
@@ -236,3 +236,55 @@ func (t *networkTool) ReadExternalDocs(ctx context.Context, args map[string]inte
 
 	return tools.ToolResult{Text: content}, nil
 }
+
+func registerNetwork(r tools.Registry, net *networkTool) error {
+	if err := r.RegisterToToolkitWithOptions("network", &tools.ToolDeclaration{
+		Name:        "read_external_docs",
+		Description: "Fetches and cleans content from a URL, stripping HTML tags and scripts to provide readable documentation. Useful for researching library APIs.",
+		Parameters: &tools.Schema{
+			Type: "OBJECT",
+			Properties: map[string]*tools.Schema{
+				"url": {
+					Type:        "STRING",
+					Description: "The documentation URL to fetch.",
+				},
+			},
+			Required: []string{"url"},
+		},
+	}, net.ReadExternalDocs, tools.ToolOptions{LongRunning: true}); err != nil {
+		return err
+	}
+
+	if err := r.RegisterToToolkitWithOptions("network", &tools.ToolDeclaration{
+		Name:        "http_request",
+		Description: "Executes a custom HTTP request.",
+		Parameters: &tools.Schema{
+			Type: "OBJECT",
+			Properties: map[string]*tools.Schema{
+				"method": {
+					Type:        "STRING",
+					Description: "HTTP method (GET, POST, PUT, DELETE, etc.).",
+				},
+				"url": {
+					Type:        "STRING",
+					Description: "The target URL.",
+				},
+				"headers": {
+					Type:        "OBJECT",
+					Description: "HTTP headers as a map of strings.",
+					Properties: map[string]*tools.Schema{
+						"Content-Type": {Type: "STRING"},
+					},
+				},
+				"body": {
+					Type:        "STRING",
+					Description: "Request body content.",
+				},
+			},
+			Required: []string{"method", "url"},
+		},
+	}, net.HttpRequest, tools.ToolOptions{LongRunning: true}); err != nil {
+		return err
+	}
+	return nil
+}
